Assume co-ordinates might use a comma as decimal separator
This restores some functionality from a previous commit where instead of
the Django template assuming that {{ some_agrifield.location.x }} is a
number using a dot as a decimal separator, it also allows a comma
instead. On production somehow there's a comma. I don't know why, but
all this shall be reworked when we move to Leaflet.